NyvoNetHunter Project | V : 4.12.8
Code structure:

    1. Added the ( is_valid_url ) function to the ( data.py ) module.


Unit testing:

    1. Added unit testing for the ( is_valid_url ) function.


file system:

    1. Added the ( test_url_validators.py ) module to the ( unit_tests ) directory.


Patches:

    1. Removed the duplications in package imports.

    2. Patched The Graphical User Interface Improper grid system.
