+++ main_program/source/ui_design/main.ui
@@ -1412,9 +1412,9 @@ QCheckBoxchecked ~ .svg-icon {
   <widget class="QLabel" name="label_4">
    <property name="geometry">
     <rect>
-     <x>70</x>
+     <x>50</x>
      <y>10</y>
-     <width>171</width>
+     <width>191</width>
      <height>31</height>
     </rect>
    </property>
@@ -1464,7 +1464,7 @@ QLabel:disabled {
 }</string>
    </property>
    <property name="text">
-    <string>Connection Status</string>
+    <string>          Connection Status</string>
    </property>
   </widget>
  </widget>
